Cast nSizeNeeded from DWORD to size_t in vrtderivedrasterband.cpp (https://trac.osgeo.org/gdal/ticket/6704).
mingw failure:

vrtderivedrasterband.cpp:471:69: error: no matching function for call to ‘min(int, long long unsigned int)’

The alternative would be to cast everything to a DWORD.


git-svn-id: https://svn.osgeo.org/gdal/trunk@36004 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtderivedrasterband.cpp
@@ -468,7 +468,9 @@ static bool LoadPythonAPI()
     EnumProcessModules(hProcess, ahModules, sizeof(ahModules),
                         &nSizeNeeded);
 
-    const int nModules = std::min(100, nSizeNeeded / sizeof(HMODULE));
+    const int nModules =
+        std::min(size_t(100),
+                 static_cast<size_t>(nSizeNeeded) / sizeof(HMODULE));
     for(int i=0;i<nModules;i++)
     {
         if( GetProcAddress(ahModules[i], "Py_SetProgramName") )
